backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sqlalchemy
sentence-transformers
faiss-cpu
PyPDF2
pandas
